refactor: merge fund history logic to use existing FundPerformance model
